Refactor AdminController to streamline endpoints and improve code organization; remove unused AuthController and related files; update .gitignore to exclude new documentation and service files; enhance Payment, Product, Service, and Staff components with hooks for better state management; clean up UI components for improved readability and maintainability.

+++ ServiceMassage/LandingPageApp.Domain/Repositories/IGenericRepository.cs
@@ -18,7 +18,7 @@ namespace LandingPageApp.Domain.Repositories
         /// Trả về null nếu không tìm thấy.
         /// Dùng FindAsync để tận dụng EF Core tracking và cache.
         /// </summary>
-        Task<T?> GetByIdAsync(int id, CancellationToken cancellation = default);
+        Task<T?> GetByIdAsync(long id, CancellationToken cancellation = default);
         Task AddAsync(T entity, CancellationToken cancellation = default);
         void Update(T entity);
         void Delete(T entity);
